Move example analyzers to separate package
