+++ openapi-validator/src/main/java/io/ballerina/openapi/validator/example/RecordFieldExampleAnalyzer.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.openapi.validator;
+package io.ballerina.openapi.validator.example;
 
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.api.symbols.AnnotationAttachmentSymbol;
